% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/hdf4.R
\name{h4read}
\alias{h4read}
\title{Read hdf4 files}
\usage{
h4read(fname, sds.name, sds, ...)
}
\arguments{
\item{fname}{Name of the hdf4 file.}

\item{sds.name}{Array containing all the SDS or vdata contained in the hdf4 file (privided by h4list).}

\item{sds}{Name of the SDS to extract.}
}
\description{
This function returns the content of a hdf4 file as a list.
}
\examples{
h4read(CALIOP.hdf,sds.CALIOP,"Latitude")
}
\keyword{hdf4}

